+++ src/models/stubs_period_verification_period_google_one_tap.rs
@@ -23,14 +23,14 @@ pub struct StubsPeriodVerificationPeriodGoogleOneTap {
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub expire_at: Option<Option<i64>>,
+    pub expire_at: Option<Option<i32>>,
     #[serde(
         rename = "attempts",
         default,
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub attempts: Option<Option<i64>>,
+    pub attempts: Option<Option<i32>>,
 }
 
 impl StubsPeriodVerificationPeriodGoogleOneTap {
