+++ src/models/stubs_period_verification_period_otp.rs
@@ -23,16 +23,16 @@ pub struct StubsPeriodVerificationPeriodOtp {
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub attempts: Option<Option<i64>>,
+    pub attempts: Option<Option<i32>>,
     #[serde(rename = "expire_at")]
-    pub expire_at: i64,
+    pub expire_at: i32,
 }
 
 impl StubsPeriodVerificationPeriodOtp {
     pub fn new(
         status: Status,
         strategy: Strategy,
-        expire_at: i64,
+        expire_at: i32,
     ) -> StubsPeriodVerificationPeriodOtp {
         StubsPeriodVerificationPeriodOtp {
             status,
